requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Search tool dependencies (optional - install as needed)
# Tavily search
tavily>=0.5.0

# SerpAPI search
google-search-results>=2.4.0

# HTTP requests (for Baidu and Metaso)
requests>=2.31.0

# MCP SDK (for advanced usage)
mcp>=1.0.0
